requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# AI/ML
langchain==0.0.350
langgraph==0.0.20
openai==1.3.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Vector DB
chromadb==0.4.21
pinecone-client==2.2.2

# Caching
redis==5.0.1

# Messaging
twilio==8.10.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
